baseline/requirements.txt
pandas==1.0.5
numpy==1.19.0
pretty_midi==0.2.9
torch==1.5.1+cu101
matplotlib==3.2.2
PyYAML==5.3.1
rich==3.2.0
scikit_learn==0.23.1
